Removed the 'download package on single result' feature Signed-off-by: Colin Shea <colin@evaryont.me>
I did mean to remove it earlier, then I attempted to work with it.
Except, I don't even *use* it, let alone like it. So, it's gone.

It was only useful before the -flags appeared, as it was really annoying
to back in your shell's history, add 'aur/' before the query,